app2/requirements.txt
Django
S3ConfigLoader
PyMySQL
mysqlclient
django-extensions
django-auth-ldap
djangorestframework
django-rest-swagger
mod-wsgi
mock==2.0.0